# docs/public/api/swagger-2.0.yaml
basePath: /api
definitions:
...
  services.AICaptureDraft:
    properties:
      items:
        items:
          $ref: '#/definitions/services.AICaptureItem'
        type: array
      warnings:
        items:
          type: string
        type: array
    type: object
...
  v1.ActionAmountResult:
    properties:
      completed:
        type: integer
    type: object
...
  v1.Results-services_AICaptureSessionOut:
    properties:
      items:
        items:
          $ref: '#/definitions/services.AICaptureSessionOut'
        type: array
    type: object
...
  v1.WipeInventoryOptions:
    properties:
      wipeLocations:
        type: boolean
      wipeMaintenance:
        type: boolean
      wipeTags:
        type: boolean
    type: object
...
  v1.aiCaptureSessionCorrection:
    properties:
      instruction:
        type: string
      revision:
        type: integer
    type: object
  v1.aiCaptureSessionCreate:
    properties:
      locationId:
        type: string
    type: object
...
  v1.aiCaptureSessionLocationUpdate:
    properties:
      locationId:
        type: string
    type: object
...
  v1.aiCaptureSessionSubmit:
    properties:
      revision:
        type: integer
    type: object
...
  validate.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        type: string
    type: object
info:
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  version: "1.0"
paths:
  /v1/actions/create-missing-thumbnails:
    post:
      description: Creates thumbnails for items that are missing them
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Create Missing Thumbnails
      tags:
      - Actions
  /v1/actions/ensure-asset-ids:
    post:
      description: Ensures all items in the database have an asset ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Ensure Asset IDs
      tags:
      - Actions
  /v1/actions/ensure-import-refs:
    post:
      description: Ensures all items in the database have an import ref
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Ensures Import Refs
      tags:
      - Actions
  /v1/actions/set-primary-photos:
    post:
      description: Sets the first photo of each item as the primary photo
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Set Primary Photos
      tags:
      - Actions
  /v1/actions/wipe-inventory:
    post:
      description: Deletes all items in the inventory
      parameters:
      - description: Wipe options
        in: body
        name: options
        schema:
          $ref: '#/definitions/v1.WipeInventoryOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Wipe Inventory
      tags:
      - Actions
  /v1/actions/zero-item-time-fields:
    post:
      description: Resets all item date fields to the beginning of the day
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.ActionAmountResult'
      security:
      - Bearer: []
      summary: Zero Out Time Fields
      tags:
      - Actions
  /v1/ai/capture/analyze:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Inventory photos (repeat field for multiple photos)
        in: formData
        name: photos
        required: true
        type: file
      - description: Selected HomeBox location UUID
        in: formData
        name: locationId
        required: true
        type: string
      - description: Correction request
        in: formData
        name: instruction
        type: string
      - description: Current AI draft JSON
        in: formData
        name: draft
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureDraft'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
      security:
      - Bearer: []
      summary: Analyze inventory photos with an OpenAI-compatible vision model
      tags:
      - AI Capture
  /v1/ai/capture/sessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Results-services_AICaptureSessionOut'
      security:
      - Bearer: []
      summary: List the current user's AI capture sessions
      tags:
      - AI Capture Sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: Selected location
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Create an AI capture session
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}:
    delete:
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete an AI capture session
      tags:
      - AI Capture Sessions
    get:
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Get an AI capture session
      tags:
      - AI Capture Sessions
    patch:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Replacement location
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionLocationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Change an AI capture session location
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/corrections:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Correction instruction
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionCorrection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Ask AI to correct a persisted capture draft
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/draft:
    put:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Draft and expected revision
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionDraftUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Save an edited AI capture draft
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/finish:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Expected uploaded count
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionFinish'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Seal an AI capture session and queue analysis
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/photos:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Normalized image
        in: formData
        name: file
        required: true
        type: file
      - description: Client idempotency UUID
        in: formData
        name: clientPhotoId
        required: true
        type: string
      - description: Capture order
        in: formData
        name: position
        required: true
        type: integer
      - description: Same-item group UUID
        in: formData
        name: captureGroupId
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.AICaptureSessionPhoto'
      security:
      - Bearer: []
      summary: Upload one photo to an AI capture session
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/photos/{photoId}:
    delete:
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Capture photo ID
        in: path
        name: photoId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete a photo from a capturing session
      tags:
      - AI Capture Sessions
    get:
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Capture photo ID
        in: path
        name: photoId
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/gif
      - image/webp
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - Bearer: []
      summary: Read an authorized AI capture session photo
      tags:
      - AI Capture Sessions
    patch:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Capture photo ID
        in: path
        name: photoId
        required: true
        type: string
      - description: Nullable same-item group UUID
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionPhotoUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionPhoto'
      security:
      - Bearer: []
      summary: Reassign or clear a photo's same-item group
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/reanalysis/{clientId}:
    delete:
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Draft client ID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Dismiss a durable capture-item reanalysis result
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/reanalyze-item:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Reviewed item and provider
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionReanalysis'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureReanalysisOut'
      security:
      - Bearer: []
      summary: Preview provider reanalysis for one reviewed capture item
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/reanalyze-items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Reviewed items and provider
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionReanalysisBatch'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Queue durable provider reanalysis for reviewed capture items
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/retry-analysis:
    post:
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Retry failed AI capture analysis
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Expected draft revision
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionSubmit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Create reviewed items from an AI capture session
      tags:
      - AI Capture Sessions
  /v1/ai/capture/sessions/{sessionId}/submit-items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Expected draft revision and selected item IDs
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/v1.aiCaptureSessionItemsSubmit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AICaptureSessionOut'
      security:
      - Bearer: []
      summary: Create selected reviewed items from an AI capture session
      tags:
      - AI Capture Sessions
  /v1/ai/items/{id}/reanalyze:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional visual-analysis guidance
        in: body
        name: payload
        schema:
          $ref: '#/definitions/v1.aiItemReanalysisRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AIItemReanalysisOut'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/validate.ErrorResponse'
      security:
      - Bearer: []
      summary: Preview Qwen reanalysis for an existing inventory item
      tags:
      - AI Capture
...
  /v1/entities/export:
    get:
      responses:
        "200":
          description: text/csv
          schema:
            type: string
      security:
      - Bearer: []
      summary: Export Entities
      tags:
      - Entities
...
schemes:
- https
- http
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"